Upd : menu create & styling css

+++ src/Controller/MenuController.php
@@ -3,19 +3,17 @@
 namespace App\Controller;
 
 use App\Entity\Menu;
-use App\Entity\Provider;
-use App\Form\Type\MenuType;
 use App\Repository\MenuRepository;
 use App\Service\AjaxService;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use App\Service\Recaptcha;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Route("/menu", name="menu_")
- * @IsGranted("IS_AUTHENTICATED_FULLY")
  */
 class MenuController extends AbstractController
 {
@@ -24,50 +22,62 @@ class MenuController extends AbstractController
      */
     public function index(MenuRepository $menuRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('menu/index.html.twig', [
             'menus' => $menuRepository->findAll(),
         ]);
     }
 
     /**
-     * @Route("/new/menu/{id}", name="new", methods={"POST"})
+     * @Route("/new/", name="new", methods={"GET", "POST"})
      */
-    public function new(Provider $provider, Request $request, AjaxService $ajaxService): Response
+    public function new(Request $request, AjaxService $ajaxService, Recaptcha $recaptcha): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_PROVIDER');
+
         $menu = new Menu();
+
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+        $provider = $user->getProvider();
         $form = $ajaxService->create_menu($menu, $provider);
 
         if ($request->isXmlHttpRequest()) {
             $form->handleRequest($request);
 
-            if ($form->isSubmitted() && $form->isValid()) {
+            $f = false;
+            if ($g = $form->get('recaptcha')->getData()) {
+                $f = $recaptcha->captchaverify($g)->success;
+            }
+            if ($form->isSubmitted() && !$f) {
+                $form->get('recaptcha')->addError(new FormError('Recaptcha : êtes-vous un robot ?'));
+            }
+            if ($form->isSubmitted() && !$form->isValid()) {
+                return $this->render(
+                    'menu/_form.html.twig',
+                    [
+                        'form' => $form->createView(),
+                    ],
+                    new Response('error', Response::HTTP_BAD_REQUEST)
+                );
+            }
+
+            if ($form->isSubmitted() && $form->isValid() && $f) {
                 $entityManager = $this->getDoctrine()->getManager();
                 $menu->setProvider($provider);
                 $entityManager->persist($menu);
                 $entityManager->flush();
 
-                return $this->render(
-                    'ajax/menu/new.html.twig',
-                    [
-                        'menu' => $menu,
-                    ],
-                    new Response('success', Response::HTTP_CREATED)
-                );
-            }
-            if ($form->isSubmitted() && !$form->isValid()) {
-                return $this->render(
-                    'ajax/menu/_form.html.twig',
-                    [
-                        'form_menu' => $form->createView(),
-                    ],
-                    new Response('error', Response::HTTP_BAD_REQUEST)
-                );
-            }
-        } else {
-            $this->addFlash('error', 'Impossible de créer un menu de cette façon !');
+                $this->addFlash('success', 'Menu créé avec success.');
 
-            return new Response('error', Response::HTTP_METHOD_NOT_ALLOWED);
+                return new Response($this->generateUrl('user_manage', ['id' => $user->getId(), 'view' => 'v-pills-menu']), Response::HTTP_CREATED);
+            }
         }
+
+        return $this->render('menu/_form.html.twig', [
+            'form' => $form->createView(),
+        ]);
     }
 
     /**
@@ -83,19 +93,46 @@ class MenuController extends AbstractController
     /**
      * @Route("/{id}/edit", name="edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Menu $menu): Response
+    public function edit(Request $request, Menu $menu, AjaxService $ajaxService, Recaptcha $recaptcha): Response
     {
-        $form = $this->createForm(MenuType::class, $menu);
-        $form->handleRequest($request);
+        $this->denyAccessUnlessGranted('MENU_EDIT', $menu);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+        $provider = $user->getProvider();
+        $form = $ajaxService->edit_menu($menu, $provider);
 
-            return $this->redirectToRoute('menu_index');
+        if ($request->isXmlHttpRequest()) {
+            $form->handleRequest($request);
+
+            $f = false;
+            if ($g = $form->get('recaptcha')->getData()) {
+                $f = $recaptcha->captchaverify($g)->success;
+            }
+            if ($form->isSubmitted() && !$f) {
+                $form->get('recaptcha')->addError(new FormError('Recaptcha : êtes-vous un robot ?'));
+            }
+            if ($form->isSubmitted() && !$form->isValid()) {
+                return $this->render(
+                    'menu/_form.html.twig',
+                    [
+                        'form' => $form->createView(),
+                    ],
+                    new Response('error', Response::HTTP_BAD_REQUEST)
+                );
+            }
+
+            if ($form->isSubmitted() && $form->isValid() && $f) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->flush();
+
+                $this->addFlash('success', 'Menu modifié avec success.');
+
+                return new Response($this->generateUrl('user_manage', ['id' => $user->getId(), 'view' => 'v-pills-menu']), Response::HTTP_CREATED);
+            }
         }
 
-        return $this->render('menu/edit.html.twig', [
-            'menu' => $menu,
+        return $this->render('menu/_form.html.twig', [
             'form' => $form->createView(),
         ]);
     }
@@ -105,12 +142,17 @@ class MenuController extends AbstractController
      */
     public function delete(Request $request, Menu $menu): Response
     {
+        $this->denyAccessUnlessGranted('MENU_DELETE', $menu);
+
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+
         if ($this->isCsrfTokenValid('delete'.$menu->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($menu);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('menu_index');
+        return $this->redirectToRoute('user_manage', ['id' => $user->getId(), 'view' => 'v-pills-menu']);
     }
 }
